+++ mailinglist/models.py
@@ -1,13 +1,19 @@
+from pathlib import Path
+
 from django.db import models
 from django.utils.timezone import now
 from django_enumfield.enum import EnumField
 from markdown import markdown
 
-from mailinglist.conf import settings
+from mailinglist.conf import hookset, settings
 from mailinglist.enum import SubmissionStatusEnum, SubscriptionStatusEnum
 
 
 class MailingList(models.Model):
+    """The fundamental model for sending messages. A mailing list can be
+    subscribed to by users and is required for composing (and therefore
+    sending) messages"""
+
     name = models.CharField(max_length=128)
     slug = models.SlugField(db_index=True, unique=True)
     email = models.EmailField(help_text="Sender e-mail")
@@ -23,15 +29,19 @@ class MailingList(models.Model):
 
     @property
     def sender_tag(self):
+        """Mailing list sender formatted in the standard way for email addresses::
+        "FirstName LastName" <email>
+        """
         return f'"{self.sender}" <{self.email}>'
 
     @property
     def published_messages(self):
+        """All messages that have been published to this mailing list prior to now"""
         return self.messages.filter(submission__published__lte=now()).distinct()
 
 
 class GlobalDeny(models.Model):
-    """Users with instances here will not receive mailing list messages"""
+    """Users with instances here will not receive mailing list messages."""
 
     user = models.OneToOneField(
         settings.MAILINGLIST_USER_MODEL,
@@ -42,6 +52,10 @@ class GlobalDeny(models.Model):
 
 
 class Subscription(models.Model):
+    """The means by which a user subscribes to receive mailing list messages.
+    **Do not create instances of this yourself, use the methods in
+    ``mailinglist.services.SubscriptionService`` for managing subscriptions!**"""
+
     user = models.ForeignKey(
         settings.MAILINGLIST_USER_MODEL,
         on_delete=models.CASCADE,
@@ -69,6 +83,10 @@ class Subscription(models.Model):
 
 
 class SubscriptionChange(models.Model):
+    """Tracks changes to subscriptions so that subscriptions/unsubscriptions
+    can be audited. Instances of this model are managed by
+    ``mailinglist.services.SubscriptionService``."""
+
     subscription = models.ForeignKey(Subscription, on_delete=models.CASCADE)
     from_status = EnumField(SubscriptionStatusEnum, null=True, blank=True)
     to_status = EnumField(SubscriptionStatusEnum)
@@ -76,6 +94,9 @@ class SubscriptionChange(models.Model):
 
 
 class Message(models.Model):
+    """The messages which will be sent to subscribed users. Comprised of
+    ``MessagePart`` and ``MessageAttachment`` instances."""
+
     title = models.CharField(max_length=128)
     slug = models.SlugField()
     mailing_list = models.ForeignKey(
@@ -83,10 +104,6 @@ class Message(models.Model):
     )
     created = models.DateTimeField(auto_now_add=True, editable=False)
     modified = models.DateTimeField(auto_now=True, editable=False)
-
-    @property
-    def subject(self):
-        return f"[{self.mailing_list}] {self.title}"
 
     def __str__(self):
         return self.title
@@ -100,6 +117,8 @@ class Message(models.Model):
 
 
 class MessagePart(models.Model):
+    """The text content of a ``Message``."""
+
     message = models.ForeignKey(
         Message, on_delete=models.CASCADE, related_name="message_parts"
     )
@@ -113,6 +132,7 @@ class MessagePart(models.Model):
         return markdown(self.text)
 
     class Meta:
+        ordering = ["order"]
         constraints = [
             models.UniqueConstraint(
                 fields=["message", "order"], name="unique_order_per_message"
@@ -120,10 +140,49 @@ class MessagePart(models.Model):
         ]
 
 
-# TODO: attachments
+def attachment_upload_to(instance, filename):
+    return Path(
+        "mailinglist-static",
+        "attachments",
+        str(instance.message.pk),
+        filename,
+    )
+
+
+def hookset_validation_wrapper(value):
+    # Model field validators are referenced in migration files. This wrapper
+    #  allows the hookset to be dynamic without causing chaos in the migration
+    #  file.
+    hookset.message_attachment_file_validator(value)
+
+
+class MessageAttachment(models.Model):
+    """The file content of a ``Message``"""
+
+    message = models.ForeignKey(
+        Message, on_delete=models.CASCADE, related_name="attachments"
+    )
+    file = models.FileField(
+        upload_to=attachment_upload_to,
+        verbose_name="attachment",
+        validators=[hookset_validation_wrapper],
+    )
+    filename = models.CharField(max_length=256, null=False)
+
+    def __str__(self):
+        return f"{self.filename} on {self.message.title}"
+
+    def save(self, **kwargs):
+        # The file name may be mangled by name collisions upon save,
+        #   capture the original filename prior to the first save.
+        if not self.pk:
+            self.filename = str(self.file)
+        super().save(**kwargs)
 
 
 class Submission(models.Model):
+    """The means by which ``Message`` instances are published and sent."""
+
     message = models.OneToOneField(
         Message, on_delete=models.PROTECT, related_name="submission"
     )
@@ -146,6 +205,10 @@ class Submission(models.Model):
 
 
 class Sending(models.Model):
+    """Tracks the sending of each ``Submission`` to each individual
+    ``Subscription``. Provided for audit, as well as to allow interruped
+    send jobs to be resumed without double-sending to any users."""
+
     submission = models.ForeignKey(Submission, on_delete=models.PROTECT)
     subscription = models.ForeignKey(Subscription, on_delete=models.PROTECT)
     sent = models.DateTimeField(auto_now_add=True)