+++ CHANGELOG.md
@@ -7,6 +7,22 @@ Changes to this project will be documented in this file.
 ### Removed
 ### Fixed
 
+
+## [0.1.3]
+### Added
+- `MessageAttachment` model and machinery to manage and send attachments.
+- `List-Unsubscribe` header to outgoing messages.
+- Lots of documentation!
+- Contents to `message_subject.txt` template.
+- Management command, celery tasks for processing published submissions.
+- View and URL for index of mailing list archives.
+- Documentation generation within CI pipeline
+### Changed
+- More fiddling with the badges
+- Message preparation pipeline to reduce template lookups
+### Removed
+- `Message.subject`
+
 ## [0.1.2]
 ### Changed
 - Readme to point to correct github workflow status